improve UX and remove some unused code
