+++ console/templates/FTP.html
@@ -8,16 +8,20 @@
 
   {% include 'partials/navigation_tabs.html' %}
 
-  {% include 'partials/receiver.html' %}
+  <div class="container">
 
-  <div class="alert hidden alert-danger" role="alert" style="overflow: auto">
-    <div class="alert hidden alert-danger panel__body" role="alert">
-      <div class="alert hidden alert-danger" role="alert">
-        {# <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">×</span></button> #}
+    {% include 'partials/receiver.html' %}
+
+    <div class="alert hidden alert-danger" role="alert" style="overflow: auto">
+      <div class="alert hidden alert-danger panel__body" role="alert">
+        <div class="alert hidden alert-danger" role="alert">
+          {# <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">×</span></button> #}
+        </div>
       </div>
     </div>
-  </div>
 
-  {% include 'partials/modal.html' %}
+    {% include 'partials/modal.html' %}
+
+  </div>
 
 {% endblock content %}
