+++ console/templates/add_user.html
@@ -8,7 +8,7 @@
 
 {% include 'partials/navigation_tabs.html' %}
 
-<div class='container'>
+<div class="container">
   <div class="row">
 
     <div class="col-sm-6">
