--- console/templates/home.html
@@ -1,24 +1,0 @@
-{% import 'partials/section.html' as section %}
-{% extends "layouts/_twocol.html" %}
-
-{% block page_title %} Home {% endblock %}
-
-
-{% block content %}
-
-{% include 'partials/navigation_tabs.html' %}
-
-<div class='container'>
-  <div class="row">
-    <div class="col-sm-6">
-      {% if current_user.is_authenticated %}
-        <h3>Logged in as: {{current_user.email}}</h3>
-      {% else %}
-        To use this service you need to
-        <a href="{{ url_for('security.login', next=request.path)}}"> log in</a>
-      {% endif %}
-    </div>
-  </div>
-</div>
-
-{% endblock content %}
