+++ console/templates/layouts/_base.html
@@ -42,6 +42,7 @@
     <link href="{{ url_for('static', filename='site.min.css') }}" rel="stylesheet" />
     <link href="{{ url_for('static', filename='css/font-awesome.min.css') }}" rel="stylesheet" />
     <script src="/static/js/jquery-3.2.1.min.js"></script>
+    <script src="/static/js/bootstrap.min.js"></script>
     <script src="/static/js/app.js"></script>
     <script src="/static/js/moment.min.js"></script>
