+++ console/templates/partials/image_data.html
@@ -1,13 +1,13 @@
-    <div class="tab-pane" id="images">
-      <table id="image-data" class="table" width="100%">
-        <thead>
-          <tr>
-            <th class="table__header-cell table__header-cell--align">Filename</th>
-            <th class="table__header-cell table__header-cell--align">Size</th>
-            <th class="table__header-cell table__header-cell--align">Modified</th>
-          </tr>
-        </thead>
-        <tbody>
-        </tbody>
-      </table>
-    </div>
+<div class="tab-pane" id="images">
+  <table id="image-data" class="table" width="100%">
+    <thead>
+      <tr>
+        <th class="table__header-cell table__header-cell--align">Filename</th>
+        <th class="table__header-cell table__header-cell--align">Size</th>
+        <th class="table__header-cell table__header-cell--align">Modified</th>
+      </tr>
+    </thead>
+    <tbody>
+    </tbody>
+  </table>
+</div>
