+++ console/templates/partials/navigation_tabs.html
@@ -1,11 +1,5 @@
 <ul class="navigation-tabs navigation-tabs--component">
 
-    <li id="Home"
-        class="navigation-tabs__item navigation-tabs__item--primary {% if request.path == '/home' %} navigation-tabs__item--active {% endif %}">
-        <a href="/home"
-           class="btn btn--secondary btn--border navigation-tabs__tab"> Home </a>
-    </li>
-
     <li id="Submit"
         class="navigation-tabs__item navigation-tabs__item--primary {% if request.path == '/submit' %} navigation-tabs__item--active {% endif %}">
         <a href="/submit"
