+++ console/templates/partials/pck_data.html
@@ -1,13 +1,13 @@
-    <div class="tab-pane" id="pcks">
-      <table id="pck-data" class="table" width="100%">
-        <thead>
-          <tr>
-            <th class="table__header-cell table__header-cell--align">Filename</th>
-            <th class="table__header-cell table__header-cell--align">Size</th>
-            <th class="table__header-cell table__header-cell--align">Modified</th>
-          </tr>
-        </thead>
-        <tbody>
-        </tbody>
-      </table>
-    </div>
+<div class="tab-pane" id="pcks">
+  <table id="pck-data" class="table" width="100%">
+    <thead>
+      <tr>
+        <th class="table__header-cell table__header-cell--align">Filename</th>
+        <th class="table__header-cell table__header-cell--align">Size</th>
+        <th class="table__header-cell table__header-cell--align">Modified</th>
+      </tr>
+    </thead>
+    <tbody>
+    </tbody>
+  </table>
+</div>
