+++ console/templates/partials/receipts_data.html
@@ -1,13 +1,13 @@
-    <div role="tab-pane" id="receipts">
-      <table id="receipt-data" class="table" width="100%">
-        <thead>
-          <tr>
-            <th class="table__header-cell table__header-cell--align">Filename</th>
-            <th class="table__header-cell table__header-cell--align">Size</th>
-            <th class="table__header-cell table__header-cell--align">Modified</th>
-          </tr>
-        </thead>
-        <tbody>
-        </tbody>
-      </table>
-    </div>
+<div class="tab-pane" id="receipts">
+  <table id="receipt-data" class="table" width="100%">
+    <thead>
+      <tr>
+        <th class="table__header-cell table__header-cell--align">Filename</th>
+        <th class="table__header-cell table__header-cell--align">Size</th>
+        <th class="table__header-cell table__header-cell--align">Modified</th>
+      </tr>
+    </thead>
+    <tbody>
+    </tbody>
+  </table>
+</div>
