+++ console/templates/partials/receiver.html
@@ -13,7 +13,7 @@
   <div id="tabs">
     <ul class="navigation-tabs navigation-tabs--component" role="tablist">
 
-      <li class="navigation-tabs__item navigation-tabs__item--primary" role="presentation">
+      <li class="navigation-tabs__item navigation-tabs__item--primary">
         <a href="#pcks"
            class="btn btn--secondary btn--border navigation-tabs__tab">PCKs</a> </li>
 
@@ -21,12 +21,12 @@
         <a href="#images"
            class="btn btn--secondary btn--border navigation-tabs__tab">Images</a> </li>
 
-      <li class="navigation-tabs__item navigation-tabs__item--primary" role="tab">
-        <a data-toggle="tab" href="#indices"
+      <li class="navigation-tabs__item navigation-tabs__item--primary">
+        <a href="#indices"
            class="btn btn--secondary btn--border navigation-tabs__tab">Indices</a> </li>
 
-      <li class="navigation-tabs__item navigation-tabs__item--primary" role="tab">
-        <a data-toggle="tab" href="#receipts"
+      <li class="navigation-tabs__item navigation-tabs__item--primary">
+        <a href="#receipts"
            class="btn btn--secondary btn--border navigation-tabs__tab">Receipts</a> </li>
     </ul>
