+++ console/templates/submit.html
@@ -8,14 +8,18 @@
 
   {% include 'partials/navigation_tabs.html' %}
 
-  {% include 'partials/submitter.html' %}
+    <div class="container">
 
-  <div class="alert hidden alert-danger" role="alert" style="overflow: auto">
-    <div class="alert hidden alert-danger panel__body" role="alert">
-      <div class="alert hidden alert-danger" role="alert">
-        {# <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">×</span></button> #}
+    {% include 'partials/submitter.html' %}
+
+      <div class="alert hidden alert-danger" role="alert" style="overflow: auto">
+        <div class="alert hidden alert-danger panel__body" role="alert">
+          <div class="alert hidden alert-danger" role="alert">
+            {# <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">×</span></button> #}
+          </div>
+        </div>
       </div>
+
     </div>
-  </div>
 
 {% endblock content %}
