+++ console/views/store.py
@@ -15,7 +15,7 @@ from structlog import wrap_logger
 from console import settings
 from console.database import db_session
 from console.models import SurveyResponse
-from console.views.home import send_data
+from console.views.submit import send_data
 from sdc.rabbit import QueuePublisher
 
 logger = wrap_logger(logging.getLogger(__name__))
