+++ tests/test_console.py
@@ -27,7 +27,7 @@ class TestConsole(unittest.TestCase):
         with mock.patch('requests.post') as mock_post:
             mock_post.return_value = r
             r.status_code = 200
-            response = views.home.send_data(logger, "", data=123)
+            response = views.submit.send_data(logger, "", data=123)
             self.assertEqual(response.status_code, 200)
 
     def test_send_data_400(self):
@@ -36,7 +36,7 @@ class TestConsole(unittest.TestCase):
             mock_post.return_value = r
             r.status_code = 400
             with self.assertRaises(ClientError):
-                views.home.send_data(logger, "", data=123)
+                views.submit.send_data(logger, "", data=123)
 
     def test_send_data_404(self):
         r = requests.Response()
@@ -44,7 +44,7 @@ class TestConsole(unittest.TestCase):
             mock_post.return_value = r
             r.status_code = 404
             with self.assertRaises(ClientError):
-                views.home.send_data(logger, "", data=123)
+                views.submit.send_data(logger, "", data=123)
 
     def test_send_data_500(self):
         r = requests.Response()
@@ -52,7 +52,7 @@ class TestConsole(unittest.TestCase):
             mock_post.return_value = r
             r.status_code = 500
             with self.assertRaises(ServiceError):
-                views.home.send_data(logger, "", data=123)
+                views.submit.send_data(logger, "", data=123)
 
 
 Postgresql = testing.postgresql.PostgresqlFactory(cache_initialized_db=False)
